Made changes for better readablity.

+++ compileapp.sh
@@ -1,4 +1,5 @@
-rm hello
+[ -f hello ] && rm hello
 g++ -std=c++11 -c Pet.cpp helloworld.cpp
 g++ -o hello Pet.o helloworld.o -lm
 rm *.o
+./hello
